Completed l-system. Basic PGTree now use l-system to make a tree.

+++ src/Vector.cpp
@@ -6,7 +6,7 @@ Vector::Vector(const Vector& orig) {
     z = orig.z;
 }
 
-double Vector::magnitude () {
+float Vector::magnitude () {
     return sqrt((x*x) + (y*y) + (z*z));
 }
